Fixed ckeditor
Had to just set to close popup after each action for now.  I had to edit
the link in your editcontent.php panel file.  Still need to find a way
to refresh the images in the view after.  the browser cache keeps a copy
of the old file and it needs to refresh from the server to work
correctly.